move CUSPARSE_TRY to `error.hpp`
