+++ cpp/include/cuspatial/utility.hpp
@@ -14,37 +14,10 @@
  * limitations under the License.
  */
 
-#include <string>
-#include <stdexcept>
 #include "cusparse.h"
 
 namespace cuspatial {
-
-struct cusparse_error : public std::runtime_error {
-  cusparse_error(std::string const& message) : std::runtime_error(message) {}
-};
-
 namespace detail {
-
-inline void throw_cusparse_error(cusparseStatus_t error,
-                                 const char* file,
-                                 unsigned int line)
-{
-  // would be nice to include `cusparseGetErrorName(error)` and
-  // `cusparseGetErrorString(error)`, but those aren't introduced until
-  // cuda 10.1 (and not in the initial release).
-  throw cuspatial::cusparse_error(
-      std::string{"CUSPARSE error encountered at: " + std::string{file} + ":" +
-                  std::to_string(line) + ": " + std::to_string(error)});
-}
-
-#define CUSPARSE_TRY(call)                                                 \
-  do {                                                                     \
-    cusparseStatus_t status = (call);                                      \
-    if (CUSPARSE_STATUS_SUCCESS != status) {                               \
-      cuspatial::detail::throw_cusparse_error(status, __FILE__, __LINE__); \
-    }                                                                      \
-  } while (0);
 
 } // detail namespace
